# okrTrackingSystem/api-docs.yaml
openapi: 3.1.0
info:
  title: OKR Tracking System API
  description: |
    ## OKR Tracking System — Garant Bank

    REST API for managing **Objectives & Key Results (OKR)**, organizational structure, user management, and multi-source performance evaluations.

    ### Key Features
    - **OKR Management** — Create and track Objectives with weighted Key Results
    - **Multi-Source Evaluations** — Director (star rating), HR (letter grade), and Business Block assessments
    - **Score Calculation** — Automatic weighted scoring: 60% OKR + 20% Director + 20% HR
    - **Organizational Hierarchy** — Divisions → Departments → Objectives → Key Results
    - **Role-Based Access Control** — 6 roles with fine-grained permissions
    - **Excel Export** — Export all OKR data to Excel spreadsheets

    ### Authentication
    All endpoints (except login and export) require a **JWT Bearer token**.
    1. Call `POST /api/auth/login` with your credentials
    2. Copy the `token` from the response
    3. Click the **Authorize** button above and enter the token

    ### Roles
    | Role | Description |
    |------|-------------|
    | `ADMIN` | Full system access, user management |
    | `DIRECTOR` | Division oversight, star-based evaluations |
    | `DEPARTMENT_LEADER` | Department OKR management |
    | `HR` | Letter-grade evaluations (A/B/C/D) |
    | `BUSINESS_BLOCK` | Business block star-based evaluations |
    | `EMPLOYEE` | View own profile and assigned departments |
  version: 1.0.0
...
servers:
...
security:
- Bearer Authentication: []
...
paths:
  /api/users/{id}:
    get:
      tags:
      - Users
      summary: Get user by ID
      description: Returns user details by ID. ADMIN and DIRECTOR can view any user.
        Other users can only view their own profile.
      operationId: getUserById
      parameters:
      - name: id
        in: path
        description: User ID (UUID)
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "403":
          description: Can only view own profile
        "404":
          description: User not found
    put:
      tags:
      - Users
      summary: Update user
      description: |-
        Update user details.

        **ADMIN** can update all fields including role, department assignments, and account status.

        **Regular users** can only update their own profile fields: fullName, email, jobTitle, phoneNumber, bio, password.
      operationId: updateUser
      parameters:
      - name: id
        in: path
        description: User ID (UUID)
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
        required: true
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "400":
          description: Invalid request
        "403":
          description: Cannot update admin-only fields
        "404":
          description: User not found
    delete:
      tags:
      - Users
      summary: Delete user
      description: Permanently delete a user account. Users cannot delete their own
        account. **Requires ADMIN role.**
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        description: User ID (UUID)
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: User deleted
        "400":
          description: Cannot delete your own account
        "403":
          description: Only ADMIN can delete users
        "404":
          description: User not found
...
  /api/users:
    get:
      tags:
      - Users
      summary: Get all users
      description: Returns all users in the system. **Requires ADMIN or DIRECTOR role.**
      operationId: getAllUsers
      responses:
        "200":
          description: List of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDTO"
        "403":
          description: Insufficient permissions
    post:
      tags:
      - Users
      summary: Create user
      description: Create a new user account with specified role and optional department
        assignments. **Requires ADMIN role.**
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
        required: true
      responses:
        "200":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "400":
          description: "Invalid request (e.g., duplicate username/email)"
        "403":
          description: Only ADMIN can create users
  /api/users/{id}/photo:
    post:
      tags:
      - Users
      summary: Upload profile photo
      description: "Upload a profile photo for a user. Accepts JPEG, PNG, and GIF\
        \ files up to 5MB. ADMIN can upload for any user; other users can only upload\
        \ their own photo."
      operationId: uploadPhoto
      parameters:
      - name: id
        in: path
        description: User ID (UUID)
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photo:
                  type: string
                  format: binary
                  description: "Profile photo file (JPEG, PNG, GIF; max 5MB)"
              required:
              - photo
      responses:
        "200":
          description: Photo uploaded — returns updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "400":
          description: Invalid file (wrong type or too large)
        "403":
          description: Can only upload own photo
  /api/users/{id}/departments:
    post:
      tags:
      - Users
      summary: Assign departments to user
      description: Assign one or more departments to a user. Replaces existing department
        assignments. **Requires ADMIN role.**
      operationId: assignDepartments
      parameters:
      - name: id
        in: path
        description: User ID (UUID)
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignDepartmentsRequest"
        required: true
      responses:
        "200":
          description: Departments assigned — returns updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "403":
          description: Only ADMIN can assign departments
        "404":
          description: User or department not found
...
  /api/users/with-scores:
    get:
      tags:
      - Users
      summary: Get all users with scores
      description: Returns all users with their overall performance scores calculated
        from assigned departments. Used for the Team Overview page. **Requires ADMIN
        or DIRECTOR role.**
      operationId: getAllUsersWithScores
      responses:
        "200":
          description: List of users with score data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserWithScoreDTO"
        "403":
          description: Insufficient permissions
  /api/users/me/profile:
    get:
      tags:
      - Users
      summary: Get my profile
      description: "Returns the extended profile of the currently authenticated user,\
        \ including detailed department information and recent evaluations."
      operationId: getMyProfile
      responses:
        "200":
          description: User profile with detailed departments and evaluations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileDTO"
        "401":
          description: Not authenticated
  /api/users/by-department/{deptId}:
    get:
      tags:
      - Users
      summary: Get users by department
      description: Returns all users assigned to a specific department.
      operationId: getUsersByDepartment
      parameters:
      - name: deptId
        in: path
        description: Department ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: List of users in the department
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDTO"
        "404":
          description: Department not found
...
  /api/auth/me:
    get:
      tags:
      - Authentication
      summary: Get current user
      description: Returns the profile information of the currently authenticated
        user based on the JWT token.
      operationId: getCurrentUser
      responses:
        "200":
          description: Current user details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "401":
          description: Not authenticated
  /api/users/{id}/departments/{deptId}:
    delete:
      tags:
      - Users
      summary: Remove department from user
      description: Remove a specific department assignment from a user. **Requires
        ADMIN role.**
      operationId: removeDepartment
      parameters:
      - name: id
        in: path
        description: User ID (UUID)
        required: true
        schema:
          type: string
          format: uuid
      - name: deptId
        in: path
        description: Department ID to remove
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Department removed — returns updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "403":
          description: Only ADMIN can remove departments
        "404":
          description: User or department not found
components:
  schemas:
    UpdateUserRequest:
      type: object
      description: Request to update user details. All fields are optional — only
        provided fields are updated. Fields marked 'ADMIN only' are ignored for non-admin
        users.
      properties:
        fullName:
          type: string
          description: Full display name
          example: John Doe Updated
        email:
          type: string
          description: Email address
          example: john.updated@garantbank.uz
        jobTitle:
          type: string
          description: Job title
          example: Lead Developer
        phoneNumber:
          type: string
          description: Phone number
          example: +998 90 987 65 43
        bio:
          type: string
          description: Bio/description
          example: Team lead with 5 years experience
        role:
          type: string
          description: User role (ADMIN only)
          enum:
          - EMPLOYEE
          - DEPARTMENT_LEADER
          - HR
          - DIRECTOR
          - BUSINESS_BLOCK
          - ADMIN
          example: DEPARTMENT_LEADER
        assignedDepartmentIds:
          type: array
          description: Department IDs to assign (ADMIN only)
          example:
          - dept-001
          - dept-002
          items:
            type: string
        password:
          type: string
          description: "New password (optional, for password change)"
          example: newSecurePass456
        isActive:
          type: boolean
          description: Account active status (ADMIN only)
          example: true
        canEditAssignedDepartments:
          type: boolean
          description: Whether user can edit assigned departments (ADMIN only)
          example: true
        readOnly:
          type: boolean
          description: Whether user is in read-only mode (ADMIN only)
          example: false
...
    UserDTO:
      type: object
      description: User information (password excluded)
      properties:
        id:
          type: string
          format: uuid
          description: User ID
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          description: Unique username
          example: john.doe
        email:
          type: string
          description: Email address
          example: john.doe@garantbank.uz
        fullName:
          type: string
          description: Full display name
          example: John Doe
        role:
          type: string
          description: User role
          enum:
          - EMPLOYEE
          - DEPARTMENT_LEADER
          - HR
          - DIRECTOR
          - BUSINESS_BLOCK
          - ADMIN
          example: EMPLOYEE
        profilePhotoUrl:
          type: string
          description: URL to user's profile photo
          example: /uploads/photo-abc123.jpg
        jobTitle:
          type: string
          description: Job title
          example: Senior Developer
        phoneNumber:
          type: string
          description: Phone number
          example: +998 90 123 45 67
        bio:
          type: string
          description: User bio/description
          example: Experienced developer focused on backend systems
        lastLogin:
          type: string
          format: date-time
          description: Timestamp of last login
          example: 2025-01-15T10:30:00
        assignedDepartments:
          type: array
          description: Departments the user is assigned to
          items:
            $ref: "#/components/schemas/DepartmentSummaryDTO"
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: 2025-01-01T09:00:00
        active:
          type: boolean
        isActive:
          type: boolean
          description: Whether the user account is active
          example: true
        canEditAssignedDepartments:
          type: boolean
          description: Whether the employee can edit their assigned departments
          example: false
        readOnly:
          type: boolean
          description: Whether the user is in read-only mode (cannot make any edits)
          example: false
...
    CreateUserRequest:
      type: object
      description: Request to create a new user (ADMIN only)
      properties:
        username:
          type: string
          description: Unique username
          example: john.doe
          minLength: 1
        email:
          type: string
          description: Email address
          example: john.doe@garantbank.uz
          minLength: 1
        password:
          type: string
          description: Password (min 6 characters)
          example: securePass123
          maxLength: 2147483647
          minLength: 6
        fullName:
          type: string
          description: Full display name
          example: John Doe
          minLength: 1
        role:
          type: string
          description: User role
          enum:
          - EMPLOYEE
          - DEPARTMENT_LEADER
          - HR
          - DIRECTOR
          - BUSINESS_BLOCK
          - ADMIN
          example: EMPLOYEE
        assignedDepartmentIds:
          type: array
          description: Optional list of department IDs to assign
          example:
          - dept-001
          items:
            type: string
        jobTitle:
          type: string
          description: Job title
          example: Senior Developer
        phoneNumber:
          type: string
          description: Phone number
          example: +998 90 123 45 67
        bio:
          type: string
          description: Bio/description
          example: Experienced backend developer
        readOnly:
          type: boolean
          description: Whether user starts in read-only mode (defaults to false)
          example: false
      required:
      - email
      - fullName
      - password
      - role
      - username
    AssignDepartmentsRequest:
      type: object
      description: Request to assign departments to a user
      properties:
        departmentIds:
          type: array
          description: List of department IDs to assign
          example:
          - dept-001
          - dept-002
          items:
            type: string
      required:
      - departmentIds
...
    UserWithScoreDTO:
      type: object
      description: User with overall performance score — used for Team Overview page
      properties:
        id:
          type: string
          format: uuid
          description: User ID
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          description: Username
          example: john.doe
        email:
          type: string
          description: Email
          example: john.doe@garantbank.uz
        fullName:
          type: string
          description: Full name
          example: John Doe
        role:
          type: string
          description: User role
          enum:
          - EMPLOYEE
          - DEPARTMENT_LEADER
          - HR
          - DIRECTOR
          - BUSINESS_BLOCK
          - ADMIN
          example: EMPLOYEE
        profilePhotoUrl:
          type: string
          description: Profile photo URL
          example: /uploads/photo-abc.jpg
        jobTitle:
          type: string
          description: Job title
          example: Senior Developer
        phoneNumber:
          type: string
          description: Phone number
          example: +998 90 123 45 67
        bio:
          type: string
          description: Bio
          example: Backend developer
        lastLogin:
          type: string
          format: date-time
          description: Last login timestamp
        assignedDepartments:
          type: array
          description: Assigned departments
          items:
            $ref: "#/components/schemas/DepartmentSummaryDTO"
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
        overallScore:
          type: number
          format: double
          description: Overall score — average of assigned department scores (0.0-1.0)
          example: 0.65
        scoreLevel:
          type: string
          description: Score level name
          enum:
          - не_соответствует
          - ниже_ожиданий
          - на_уровне_ожиданий
          - превышает_ожидания
          - исключительно
          example: на_уровне_ожиданий
        scoreColor:
          type: string
          description: Color associated with score level
          example: "#4CAF50"
        scorePercentage:
          type: number
          format: double
          description: Score as percentage (0-100)
          example: 75.0
        active:
          type: boolean
        isActive:
          type: boolean
          description: Account active status
          example: true
        canEditAssignedDepartments:
          type: boolean
          description: Can edit assigned departments
          example: false
        readOnly:
          type: boolean
          description: Whether the user is in read-only mode
          example: false
    UserProfileDTO:
      type: object
      description: Extended user profile with detailed department info and recent
        evaluations
      properties:
        id:
          type: string
          format: uuid
          description: User ID
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          description: Username
          example: john.doe
        email:
          type: string
          description: Email
          example: john.doe@garantbank.uz
        fullName:
          type: string
          description: Full name
          example: John Doe
        role:
          type: string
          description: User role
          enum:
          - EMPLOYEE
          - DEPARTMENT_LEADER
          - HR
          - DIRECTOR
          - BUSINESS_BLOCK
          - ADMIN
          example: EMPLOYEE
        profilePhotoUrl:
          type: string
          description: Profile photo URL
          example: /uploads/photo-abc.jpg
        jobTitle:
          type: string
          description: Job title
          example: Senior Developer
        phoneNumber:
          type: string
          description: Phone number
          example: +998 90 123 45 67
        bio:
          type: string
          description: Bio
          example: Backend developer
        lastLogin:
          type: string
          format: date-time
          description: Last login timestamp
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
        assignedDepartments:
          type: array
          description: Summary list of assigned departments
          items:
            $ref: "#/components/schemas/DepartmentSummaryDTO"
        assignedDepartmentsDetail:
          type: array
          description: Full department details including objectives and scores
          items:
            $ref: "#/components/schemas/DepartmentDTO"
        recentEvaluations:
          type: array
          description: Recent evaluations for this user
          items:
            $ref: "#/components/schemas/EvaluationDTO"
        active:
          type: boolean
        isActive:
          type: boolean
          description: Account active status
          example: true
        readOnly:
          type: boolean
          description: Whether the user is in read-only mode
          example: false
...
  securitySchemes:
    Bearer Authentication:
      type: http
      description: Enter the JWT token obtained from `POST /api/auth/login`
      name: Bearer Authentication
      scheme: bearer
      bearerFormat: JWT